days051-060/day057/Finify/requirements.txt
finnhub-python
python-dotenv